Break up the examples in examples/subusers/subusers.java to their own files
