+++ examples/subusers/EnableOrDisableUser.java
@@ -1,0 +1,31 @@
+import com.fasterxml.jackson.databind.JsonNode;
+import com.fasterxml.jackson.databind.ObjectMapper;
+
+import com.sendgrid.*;
+
+import java.io.IOException;
+import java.util.HashMap;
+import java.util.Map;
+
+//////////////////////////////////////////////////////////////////
+// Enable/disable a subuser
+// PATCH /subusers/{subuser_name}
+
+
+public class EnableOrDisableUser {
+  public static void main(String[] args) throws IOException {
+    try {
+      SendGrid sg = new SendGrid(System.getenv("SENDGRID_API_KEY"));
+      Request request = new Request();
+      request.setMethod(Method.PATCH);
+      request.setEndpoint("subusers/{subuser_name}");
+      request.setBody("{\"disabled\":false}");
+      Response response = sg.api(request);
+      System.out.println(response.getStatusCode());
+      System.out.println(response.getBody());
+      System.out.println(response.getHeaders());
+    } catch (IOException ex) {
+      throw ex;
+    }
+  }
+}
