+++ examples/subusers/RetrieveMonthlyEmailStatistics.java
@@ -1,0 +1,36 @@
+import com.fasterxml.jackson.databind.JsonNode;
+import com.fasterxml.jackson.databind.ObjectMapper;
+
+import com.sendgrid.*;
+
+import java.io.IOException;
+import java.util.HashMap;
+import java.util.Map;
+
+
+//////////////////////////////////////////////////////////////////
+// Retrieve the monthly email statistics for a single subuser
+// GET /subusers/{subuser_name}/stats/monthly
+
+
+public class RetrieveMonthlyEmailStatistics {
+  public static void main(String[] args) throws IOException {
+    try {
+      SendGrid sg = new SendGrid(System.getenv("SENDGRID_API_KEY"));
+      Request request = new Request();
+      request.setMethod(Method.GET);
+      request.setEndpoint("subusers/{subuser_name}/stats/monthly");
+      request.addQueryParam("date", "test_string");
+      request.addQueryParam("sort_by_direction", "asc");
+      request.addQueryParam("limit", "1");
+      request.addQueryParam("sort_by_metric", "test_string");
+      request.addQueryParam("offset", "1");
+      Response response = sg.api(request);
+      System.out.println(response.getStatusCode());
+      System.out.println(response.getBody());
+      System.out.println(response.getHeaders());
+    } catch (IOException ex) {
+      throw ex;
+    }
+  }
+}
